requirements.txt
apispec==6.3.0
awscli==1.27.137
blinker==1.6.2
boto3==1.26.135
botocore==1.29.135
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docutils==0.16
dynamodbgeo==0.0.3
dynamorm==0.11.0
Flask==2.3.2
Flask-Compress==1.13
Flask-Cors==3.0.10
flask-smorest==0.42.0
future==0.18.3
geographiclib==2.0
Geohash==1.0
geopy==2.3.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
pyasn1==0.5.0
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.30.0
rsa==4.7.2
s2sphere==0.2.5
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.15
webargs==8.2.0
Werkzeug==2.3.4
zipp==3.15.0
